refactor code, fix bug in shrink state machine payload
